Update future information and fix .good
This was turned into a future with the associative domain literals
change.  A bit of further investigation shows that the failure is due
associative domains not supporting array index types.

FWIW, the initial intent (and result) of the test was possibly lost
with the introduction of associative domain literals (coupled with
array literals).  I'm not sure how meaningful this test is anymore,
but I've updated the output and future description to reflect the
change.



git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@21218 3a8e244f-b0f2-452b-bcba-4c88e055c3ca