Fixing bug with extern classes
[reviewed by sungeun]

This commit fixes a bug that I inadvertantly introduced in r21075 when
I changed the names of the structures generated by the compiler for
classes.  For external classes, I accidentally broke the strings that
were code generated for the malloc()s and ended up malloc-ing the size
of a pointer rather than the struct representing the object.  None of
our tests picked this up (so far as I can tell) because our extern
class tests all use classes that are no bigger than a pointer in size.

To prevent this mistake from occurring again, I'm checking in a test
that uses a bigger class which results in memory problems (under
valgrind and in my executions also outside of valgrind) should it
happen again.

One note to point out is that, traditionally, we have required an
external class named "myclass" to name its struct "_myclass".  In
r21075, I changed the compiler-generated name for such a class to be
chpl_myclass_object (a name that Michael and I agreed upon at the
time).  In this commit, I originally thought I'd unify these names for
consistency, but it seems dumb to require the user class to start with
a chpl_ prefix, and verbose/ugly to force them to add a _object suffix
(something that I hadn't considered in the original commit).  So
rather than requiring them to unify the name with internal Chapel
practices, I kept the naming requirements the same as in previous
releases for simplicity.  I think it's worth considering using a
different name -- preferably one without a leading underscore -- but
left that conversation/decision for another day.




git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@21222 3a8e244f-b0f2-452b-bcba-4c88e055c3ca