Updates to README.extern, particularly to describe extern blocks and C aliases
[co-developed/reviewed with hilde]

The main goal of this commit was to add a description of extern blocks
to README.extern, which we did.  At present, it's at the end of the
file, owing to its new-ness as a feature.  Ultimately, it might
arguably be moved to the top of the file, as its use (when it works)
may render traditional extern declarations as being a more
manual/painstaking approach.

While here, did some general wordsmithing, reorganization, framing.
The biggest such change was probably the mention of the c_* type
alises defined in StandardCTypes.chpl -- a good catch on Tom's part.



git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@21226 3a8e244f-b0f2-452b-bcba-4c88e055c3ca