Minor spec edits related to tuples.
Reviewed and contributed to by Brad.

* The Procedures chapter defines 'proc tuple(x...)' to illustrate varargs.
Even though (or perhaps "because") there is no more tuple() in the
modules, this example is fine and instructive. We improved the explanation.

* The Tuples chapter put a space after the comma in 1-tuple literals.
I removed it because the space author is fine either way, but the no-space
version is more common in our modules and seems to be what Python prefers.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@21238 3a8e244f-b0f2-452b-bcba-4c88e055c3ca