Changed from [t]csh to bash syntax for consistency
Changed all READMEs to use bash syntax (when they didn't already)
and generally removed text that indicated which shell was being
used (for simplicity).  Added a top-level note to README.chplenv
stating that we use bash and giving a quick conversion to csh/tcsh
to cover ourselves elsewhere.

Added a few missing lines to README.cray that seem to have been
inadvertantly dropped along the way.



git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@21240 3a8e244f-b0f2-452b-bcba-4c88e055c3ca