Spec changes: * adjusted the spec's set of keywords * made the underscore an identifier and a keyword * allowed the underscore in the tuple value syntax
Co-developed with Brad.

The underscore became a keyword in r20146, where it was (re-)introduced
as a means of dropping tuple components on the floor while detupling.
We adjust the keyword list and the tuple value syntax appropriately.
We make the underscore an identifier because it feels right.

Keyword notes:

* in the Keywords section of Lexical_Structure.tex:

 add: _ align export extern inline

 remove: true false locale opaque range

BTW pragma, __primitive, __primitive_loop are keywords in our parser, but
not in the language spec - because they are our implementation internals.

The change "by" -> "otherwise" is in the "invisible" environment.
I think it restricts (from below) the column size, so I made it
uniform with the other columns.

* in chapel_listing.tex (for the Chapel environment):

 add: export lambda

 remove: tuple

 reorder alphabetically: int <-> inout

 todo: make the underscore also be bold-ified as a keyword

 historically, we have also been bold-ifying the following identifiers -
 these are unaffected by this change:
   bool true false int uint real imag complex string locale range opaque


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@21259 3a8e244f-b0f2-452b-bcba-4c88e055c3ca