Eliminated duplication in start_test for generating performance graphs via createGraphs for top level .graph files.
[reviewed by Sung]

This special-purpose code was added in -r18861 to ensure performance
graphs get generated for memleaks (and perhaps spectests, too). That's
because we have memleaks.graph and spectests.graph, but not the
corresponding .chpl files, in $CHPL_HOME/test.

However, start_test has general-purpose code that looks for .graph
files in each directory it processes. So when start_test is invoked on
$CHPL_HOME/test (which in our current usage it is), those .graph files
are found and processed by this general-purpose code.

As it currently stands, the special-purpose code results in duplicate
processing of top-level .graph files, so we remove it. 


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@21268 3a8e244f-b0f2-452b-bcba-4c88e055c3ca