Initial implementation of Locale-Aware Memory Management.
Pipes memory requests for objects created by Chapel through chpl_here_alloc().
That function dispatches memory requests to a locale-specific alloc() routine.
The alloc() routine is dynamically dispatched, using "here" as the "this"
pointer.  In that way, memory requests are made locale-aware.

The material changes are on line 666 of scopeResolve.cpp (to send requests to
chpl_here_alloc() and in the ChapelLocale.chpl module (to implement
chpl_here_alloc() and chpl_here_free()).  The remaining changes were massaging
to get the compiler to work with that new idea.

A significant shortcoming of this patch is that it does not replace all
calls to PRIM_CHPL_ALLOC with calls to chpl_here_alloc().  The result is that
some chunks of memory will be allocated through locale specific allocators and
others will not.  To get to release quality, these allocations will all have to
be converted to use chpl_here_alloc() (or reasons given why this is not done).


DETAILS
=======

    compiler/include/primitive.h
    compiler/AST/primitive.cpp
    compiler/AST/expr.cpp
    compiler/passes/parallel.cpp
Renamed PRIM_CHPL_ALLOC and PRIM_CHPL_FREE as PRIM_CHPL_MEM_ALLOC and
PRIM_CHPL_MEM_FREE, to reflect their actual implementation in the runtime.

Added PRIM_SIZEOF.  This is needed because locale.alloc() cannot be
simultaneously generic and dynamic and it needs to be dynamic.  Not being
generic means that we can't just pass in the type of the thing being allocated
(that would be generic), we have to pass in its size instead.

Added PRIM_CHPL_MALLOC and PRIM_CHPL_FREE.  I thought I might need these for
direct calls but probably not, so they can be removed before checkin.


    modules/internal/ChapelLocale.chpl
Added default implementations for locale.alloc() and locale.free().  These
should probably go in DefaultArchitecture.chpl and the implementations here made
"pure virtual".

Typo "construct" -> "contructor".
Added chpl_here_alloc() and chpl_here_free().


    modules/internal/ChapelSyncvar.chpl
Made _syncvar and _singlevar inherit from "object", so they can be allocated
through chpl_here_alloc().  This involved removing the "no object" pragma.

Add _cast routines, to alloc sync and single vars to be explicitly cast to
"object" so they can be freed.


    modules/internal/ChapelSyncvar.chpl
    modules/internal/ChapelBase.chpl
When chpl__maybeAutoDestroyed() is called with a variable of type sync(foo), the
sync(foo) -> foo coercion is applied and then argument type resolution is run
again.  This causes chpl__maybeAutoDestroyed(sync(int)) to be resolved as
chpl__maybeAutoDestroyed(object) -- which I consider correct.  However, for
reasons that are ill-understood (by me at least), this causes the mapping 
chpl__maybeAutoDestroyed(int) --> chpl__maybeAutoDestroyed(object) to be cached
in functionResolution, and causes a type mismatch when the case
chpl__maybeAutoDestroyed(int) really comes up.

Anyway, my workaround short-circuits that problem by getting an exact match with
_syncvar and _singlevar.

In ChapelBase, I rewrote chpl__maybeAutoDestroyed() to return true by default,
and then added the numeric case to pick up both imag and complex types (as well
as real which was not included in the original).


    compiler/include/scopeResolve.h
    compiler/passes/scopeResolve.cpp
Factored out add_root_type() and made it public, so it can be called from
functionResolution.
Made skipSet local to scopeResolve().


    compiler/AST/iterator.cpp
Recognized that code to add the base object is redundant in this file and
removed it.


    compiler/passes/normalize.cpp
Replaced PRIM_CHPL_ALLOC with calls to "chpl_here_alloc".
Added comments about how normalization hides the declared return type in the
initializer field.  


    compiler/passes/parallel.cpp
Note that some of the allocation done to support passing argument bundles
etc. across the network is just using the generic allocator.  This is because
the parallel pass is run after resolution, so I expect we have to do some tricky
things with generic expansion and resolution to call chpl_here_alloc instead.
Factored out and commented getHeapVars().


    compiler/resolution/generics.cpp
Added special case and accompanying comments in getNewSubType() so that sync and
single variables are *not* coerced to their defining types when chpl_here_alloc
and chpl_here_free are called.  This is more special-case code that would go
away if we had implicit casts and syncvars were just plain-old classes.


    compiler/resolution/wrappers.cpp
Added more special-case code to deal with freeing a dynamically-allocated sync
or single variable, with explanatory comments.


    compiler/resolution/lowerIterators.cpp
The call to chpl_here_alloc will have to be reverted to PRIM_CHPL_MEM_ALLOC,
because lowerIterators occurs after resolution, so this function call won't get
resolved (in the very few cases where it actually makes it into the generated
code).


    compiler/resolution/functionResolution.cpp
Make iterator classes derive from dtObject so they can be freed.
4857:Remove the parent class from the iterator class so it can be re-added?
This seems pretty kludgy.  It works for now, but we should perhaps find a better way.
5302:Outdented and commented code for resolving initializers.


    test/users/vass/type-tests.good
Turning sync and single into true classes changed the results of this test.


git-svn-id: http://svn.code.sf.net/p/chapel/code/branches/collaborations/hierarchical_locales@21276 3a8e244f-b0f2-452b-bcba-4c88e055c3ca