Fix for the PGI bug related to struct initializer syntax. [reviewed by ]
The PGI compiler (13.3.0) provides only a partial implementation of designated
initializers, as specified in the C99 standard.  

PGI allows casting the initializer as a whole to a type convertible to the type
of the variable being initialized (this is the <assignment-expression> syntax),
and allows supplying designators for the elements in an initializer, but it does
not support the <assignment-expression> syntax as an element within an
initializer list.  The net effect is that initializers must be "flat" -- meaning
that all of the elements in PGI-supported initializers must be of fundamental
type.

To initialize structs that contain other structs, the workaround provided by
this patch is to create a temporary of the destination type and then initialize
fields within that type using separate assignment statements for each.  For
leaf-level initializers, it uses the position-dependent initializer syntax to
create an anonymous struct.  We could use a temp and field assignments to
achieve compatibility with older C compilers, as noted in the comments.

DETAILS:

    compiler/include/expr.h
createTempVar() was converted to a static.

    compiler/AST/expr.cpp
createTempVar() was renamed as createTempRef(), to more closely match its actual use.
A new function createTempVar() was added, to create a temporary variable (not a ref).
In codegenLocaleID(), removed the cast to destination type and field designators.
createTempVar() is like createTempRef(), except that it does not set ret.isLVPtr
= GEN_PTR, and it does not add the address-of '&' operator to the variable name.

For creating wide pointers from components during code generation, we use a temp
and field assignments.  This gets around the PGI bug, which is triggered when
trying to assign a struct (chpl_localeID_t) to a field in an initializer.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@21286 3a8e244f-b0f2-452b-bcba-4c88e055c3ca