Added a test with a 'begin' within a 'forall', whose leader iterator has multiple yields.
At present the compiler in this case replicates the body of the forall
for each of the yields in the leader. This results, e.g., in creation of
multiple begin_fn functions, wrappers, and arg bundle classes.

The array is here to test correctness of autocopies/destroys.
Just two yields within the leader would be sufficient to exercise the above,
but it seems an 'if' is also needed. A couple of extra yields are just frills.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@21289 3a8e244f-b0f2-452b-bcba-4c88e055c3ca