Added a new linker flag -no_template_limits to avoid cray-mta linker errors
When building the Chapel compiler with the cray-mta compiler, we've
been getting a failure on the first attempt each time due to an
assumption in EDG that the number of templates we're instantiating
suggests an infinite loop.  Mike R. added a new flag to override this
behavior.  This commit turns this flag on when linking using C++.
I've also changed our .dotfiles to use the OKAYNIGHT version of the
compiler rather than REL/NEXT version.  It may take a day or two
for the OKAYNIGHT version to be updated with the new flag, but once
it does, the nightly failures related to building the compiler with
cray-mta should go away.



git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@21291 3a8e244f-b0f2-452b-bcba-4c88e055c3ca