Adding support for small-int-idxType ranges/domains
[reviewed by diten]

With the newly-added support for operators on small-int combinations,
we can now trivially add support for ranges and rectangular domains
that have idxTypes other than int(32)/int(64).  (We could have done it
previously, but it would have required inserting a lot of downcasts on
all the built-in operators on ranges and domains.

This commit extends the build_range() overloads to support int/uint
idxTypes of arbitrary size.  As an added bonus, it also gets rid of a
'hilde sez' comment which proved to be incorrect/outdated.  The only
other changes are to existing tests to update behavior as a result of
this change.  My approach was to update .goods where possible and to
move illegal cases out into new *TypeError2.chpl tests to mirror what
had been done after a previous change.

As an example of something that would have been illegal and is now
illegal, consider:

         var i8: int(8) = 1;
         var j8: int(8) = 10;
         var r = i8..j8;

         var k32 = int(32) = 2;

         ...r by 2...

Prior to this commit, r would've been a range with idxType of int(32)
since that was the smallest idxType that we supported.  Applying an
int(32) stride to an int(32) range is legal so this would've been
legal.  Under this change, r's idxType is int(8) and an int(32) stride
is currently not permitted for it, so it becomes an error.

I added a new test (smallIntDom.chpl) to show that the idxType of
ranges and domains defined using a small int size was correct and that
iterators over such ranges/domains also yielded the smaller int size.




git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@21292 3a8e244f-b0f2-452b-bcba-4c88e055c3ca