Adding/updating futures related to param mismatches
In adding the previous commit that supported small-int ranges and
domains, David and I were puzzling over why Chapel's rules about param
coercion only applied to default-sized ints (I think the short answer
is "because that's what C# did", but that doesn't seem sufficiently
compelling today).

Here, I'm filing a new future (smallIntParamCoerce) to ask the
question whether we should do similar parameter coercion for other int
sizes.  I'm also updating a pre-existing future (paramMismatch) that
was requesting better error messages for param actuals that failed to
resolve but for a case that now works (presumably due to improvements
to function resolution over the past year).  I updated the test to
fail again and re-capture the original behavior.

Note that only one of these futures can be made to pass ultimately --
if we decide the coercion is legal, the paramMismatch future becomes
invalid (or else has to be recast to make sense for some other
types/sizes); if we decide it's illegal, the smallIntParamCoerce test
can be retired.



git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@21293 3a8e244f-b0f2-452b-bcba-4c88e055c3ca