Initial support for .bad files in the testing system
This commit adds a capability to the testing system in which a .future test
can have a .bad file to lock in the bad behavior that the test is currently
generating.  The goal of doing so is to identify changes that cause the test
behavior to change in a way that causes the test to still not succeed but
for the wrong reasons.  In the past, examples of this have been things like
not updating the syntax of a .future test.  Thus, the test would continue
failing, but for the wrong reasons (e.g., "Syntax error" rather than
whatever the .future was intended to test).  By creating a .bad file, the
syntax error would become known and the .future could be updated.

Note that the creation of a .bad file with a .future is completely optional,
but recommended to keep the future viable.  If the future's output is
unpredictable in its incorrectness then a .bad file is clearly not a good
idea.

At present, my approach is fairly simple:  If a foo.future test fails to
pass, the test system prints out the normal message and then diffs the
output against the .bad file if it exists.  If the diff is clean, it prints
out a message

        [Clean match against .bad file for ...]

(Note that I intentionally did not use "Success" in order to avoid getting
credit for simply getting bad output).  If the diff is wrong, it prints out:

        [Error matching .bad file for ...]

My thought in making this a full-fledged error (rather than a
Future-annotated Error) is that it should get attention in short order. And
the only other way I could think to do this was to create a whole new
category of error types to diff against and summarize which felt like
overkill if this is sufficient.

Interestingly, this change to sub_test shows that the two .bad files I
committed today were incorrect due to failure to update them on my part
after initially creating them and then further modifying the test.

A possible future direction for this change:

* Currently there is only one .bad file per test; over time we may want
  to introduce increasingly complex .bad files (for platform, comm layer,
  etc.).  My thinking at present is that most futures probably fail in
  a single way, or that skipifs could be used to only test it in a certain
  case if necessary (or the test could simply choose not to have a .bad
  file).  If we did this, we should refactor the code that creates the
  .good filename and reuse it to create the .bad filename as well.




git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@21295 3a8e244f-b0f2-452b-bcba-4c88e055c3ca