Some code cleanup and commentary related to Chapel memory allocation. [reviewer: vass]
    compiler/passes/normalize.cpp

Comments were added before checkUseBeforeDefs() indicating this this check
relies on placing initialization code in the tree which (at least in the case of
classes) is redundant with guaranteed initialization.  It is also noted that the
effect of ref, out, and inout arguments on the state (defined or not) of a
variable in the calling routine is not taken into account.  This leads to excess
initializations, and an artificial restriction on the structure of the AST
(which may inhibit other features).
The claim is made that this test is being performed too early in the compilation
process, and that it should therefore be removed.

Some comments were added to the body of checkUseBeforeDefs(), and the if
... else if ... structure changed to a sequence of "if" clauses that are
executed unconditionally.  The new structure is easier to understand, because it
does not require the reader to bear in mind the negative cases.  It does not
require that the cases be mutually exclusive; it would function correctly even
if they were not.


    compiler/passes/scopeResolve.cpp
Factored out resolveUnresolvedSymExprs().  Turned the skipSet from a static to a
local variable, since its lifetime does not extend outside that new routine.

    compiler/passes/scopeResolve.cpp
    compiler/resolution/generics.cpp
Turned an alist.add() calls into add_exclusive() and assert that the result is
true, so we are sure each such addition is unique.

    compiler/passes/parallel.cpp
Factored out and commented getHeapVars().  Added an assert that its caller is
not called when fLocal is true (based on inspection).

    compiler/resolution/functionResolution.cpp
Factored conditionals for doing primitive initialization.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@21297 3a8e244f-b0f2-452b-bcba-4c88e055c3ca