Replace (most) strcmp calls for detecting chpl__autoCopy, chpl__autoDestroy, and chpl__initCopy functions with flag check
[reviewed by hilde]

This commit adds new pragmas/flags for chpl__autoCopy,
chpl__autoDestroy, and chpl_initCopy functions (FLAG_AUTO_COPY_FN,
FLAG_AUTO_DESTROY_FN, and FLAG_INIT_COPY_FN).  These pragmas are
required on the functions, and the resultant flags are used during
compilation in lieu of strcmps.  In support of my updated reference
counting work and future RAII work, I also added FLAG_DONOR_FN to
indicate a function that gives up ownership of the data being returned
(that is, the calling function is responsible for deletion).  This
pragma is required of all chpl__autoCopy functions, but should not be
put on any user functions yet.  More guidance on how to use the pragma
will be established during the RAII work.

Note that I did not remove the calls to isNamed (which does a strcmp)
for these functions, because those can (and do) take unresolved
symexprs.

Compiler changes:

- compiler/include/flags_list.h: added new flags

- Replaced occurrences of strcmp with the above function names with a
  flag check throughout

- compiler/AST/symbol.cpp: copy flags in FnSymbol::copyInner

- compiler/resolution/functionResolution.cpp: better commented
  requiresImplicitDestroy() and the use of FLAG_DONOR_FN here.

Module changes:

- Added new pragmas to the above functions

Test changes:

- test/users/ferguson/recordbug3.chpl defines a chpl__initCopy, which
  now requires the pragma.



git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@21300 3a8e244f-b0f2-452b-bcba-4c88e055c3ca