Make all non-exported symbols "static" in generated C code
[reviewed by diten; and hilde co-reviewed the first draft]

This is a change we've talked about for years but, to my surprise,
never implemented.  It essentially marks all symbols in the generated
code 'static' unless they are 'export'ed.  The goal here is to provide
as much information to the C compiler as possible for its use in
optimization; and to minimize the bleeding of Chapel-generated symbols
into the C space as much as possible.

This approach obviously relies on our current "all Chapel code goes
into a (conceptially) monolithic C file (via #includes)" and wouldn't
work if we went down the path of creating a separate .o file per
Chapel module.  However, even if and when we do implement that change
(proposed primarily for compilation speed benefits), I propose that we
will still want to maintain a monolithic compilation mode for the
purposes of maximizing information to the C compiler -- think of it as
an "optimize for compilation time vs. execution time" flag (asssuming
our assumptions are correct).  So this change seems worthwhile even
in the long-term if we move away from monolithic codegen by default.

The approach I took here was to manually mark things that are referred
to in the C runtime as being 'export', either via compiler changes or
changes to the internal modules.  (For those who haven't used it,
'export' is our way of making a Chapel symbol available for calling
from within C).

I added a new developer pragma/flag named FLAG_EXPORT_INIT which
specifies that a module's initialization routine should be exported.
There was some discussion about using a module-level export flag for
this purpose, but I was concerned that this could send the wrong
message -- implying, for example, that perhaps all of the module's
symbols should be exported.

It's worth noting that a number of the symbols that I'm exporting
should not need to be exported as we normalize the initialization of
the threads layer and runtime (work in progress on the hierarchical
locales branch).  ChapelThreads' initialization function is one
example; config consts that are referred to within the runtime are the
remaining cases that I'm aware of.

Another change that may need to occur in the future is to export the
module initialization functions of any modules that themselves export
symbols -- I don't think we've clarified the question of module
initialization initiated from the C code side of things (certainly
no tests exercise it), so I didn't tackle that here.

We talked about adding a developer flag to disable these 'static'
tags, but decided to punt on it for now under the argument that
nobody anticipated a need for it immediately and that it would
be no harder to add later.




git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@21302 3a8e244f-b0f2-452b-bcba-4c88e055c3ca