Remove GPU code from the compiler, runtime, Makefiles
[reviewed by hilde]

This commit removes the GPU code from the Chapel compiler, runtime,
and Makefiles for the following reasons:  

(1) it hasn't been tested in ages and may not work;

(2) the most recent GPU work has been done on
branch/collaborations/gpu, and that's the place to look for GPU
execution at present rather than on trunk.  Merging that code into a
GPU-less trunk would probably be easier than merging it into the
GPU-ful trunk;

(3) in our next step for GPUs, we want to integrate the emerging
hierarchical locales work, which will suggest taking a fairly
different approach.

To this end, removing GPU code from trunk and referring to the branch
when wanting to see how things were done in Albert's work seems like
the best way to move forward.


I essentially undertook this by culling a number of GPU-related
symbols and conditionals based on those symbols:

* the primitives PRIM_GPU_GET_ARRAY, PRIM_GPU_GET_VALUE,
  PRIM_GPU_GET_VAL, PRIM_GPU_ALLOC, PRIM_COPY_HOST_GPU,
  PRIM_COPY_GPU_HOST, PRIM_GPU_FREE, and PRIM_ON_GPU

* the flags FLAG_GPU_ON, FLAG_GPU_CALL

* filenames containing 'gpu'

* the --gpu flag

* the gpuFlattenArgs() pass

* the ENABLE_GPU #define

* the canCopyFromDevice/Host hooks used to specify CPU-to-GPU
  transfers in module code

* the gpusrcfile argument to codegen_makefile

* Makefile rules specific to NVCC



git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@21311 3a8e244f-b0f2-452b-bcba-4c88e055c3ca