Switched a variable shared between two non-synchronizing tasks from plain to atomic, in parallel/taskPar/figueroa/taskParallel.chpl.
[discussed within the Chapel group]

The basic code pattern is:

  begin {
    while notThereYet {} // spin
  }
  begin {
    notThereYet = true;
  }

In the absence of synchronization, the update in the second task
is not guaranteed to be visible in the first task.

To synchronize, one can access a shared synchronization variable
(which currently in Chapel causes a memory fence) in both tasks,
make notThereYet itself a sync variable or make it an atomic variable.

Since there is already an atomic variable, I opted to use an atomic.

Merging the first notThereYet.write(true) into notThereYet's
declaration would not work. That's because currently Chapel treats
initialization as a declaration followed by an assignment, and atomics
cannot be assigned directly.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@21312 3a8e244f-b0f2-452b-bcba-4c88e055c3ca