Don't depend on internal error line numbers, or the compiler version.
[Reviewed by Brad.]

The .bad file was trying to match the entire error message for an
internal error, and failed because the compiler version didn't match
exactly.  In addition, it would have failed if the line number in the
compiler at which the internal error was produced changed.  Take both of
these out of consideration by making the .bad just match the message's
initial "error: internal failure" string.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@21316 3a8e244f-b0f2-452b-bcba-4c88e055c3ca