In the PRIM_MOVE case in CallExpr::codegen(), replace chained if ... else if ... clauses with a switch [trivial]
The existing code is hard to debug, because it requires stepping down through each clause,
or setting a breakpoint on the expected clause.  Recoding the logic as a switch should make
debugging easier.

The code is structured as a number of special cases to be handled if the RHS is a primitive, 
with the possibility that unhandled cases will fall through to the general PRIM_MOVE
implementation below.  This control flow is duplicated in the switch by initially assuming
that all RHS primitives will be handled specially (handled = true) and then specifically
marking the unhandled cases (handled = false) so that the existing fallthrough behavior
is preserved.  

If handled == true at the bottom of the inner switch statement, then control reaches a break
there which exits the outer switch statement as well -- signalling that processing of the
PRIM_MOVE primitive is complete.

As a slight cleanup, three identical cases for PRIM_GET_MEMBER were consolidated into one.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@21325 3a8e244f-b0f2-452b-bcba-4c88e055c3ca