[Reviewed by Brad.]
During the r21314 review Brad noticed that chpl_mem_actualSharedHeap()
seemed not to be used.  I confirmed that this was so, and am removing it
and everything else that was only present to support it.

I added chpl_mem_actual_shared_heap() (its original name) in r18963, but
apparently only as a hedge against future needs, because I cannot find
any evidence that it was ever used.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@21326 3a8e244f-b0f2-452b-bcba-4c88e055c3ca