Enable stack overflow detection in qthreads.
[Reviewed by Sung.]

This corresponds to the following changes in the public qthreads
repository:
  7821c44918799882095ef68008bdb1e6edaadb19
  919a7fb869ef8b6506da76e4c1f46815dec93030
  9fbde8393b99702a927bc607f1574570677ecda7

Overflowing (and underflowing) references are detected by placing
untouchable guard pages above and below the stack space.  This will
cause SIGSEGV signals when such references are made.  This doesn't
allow an immediate and unambiguous diagnosis when stack overflow
occurs, because all kinds of addressing problems can cause SIGSEGV.
But, at least it prevents wild loads and stores and always produces
a nonzero exit code.

With this commit we enable the guard page protection by adding the
--enable-guard-pages option during configuration.  Then, it can be
turned on and off by setting QT_GUARD_PAGES to 1/yes/true (which is
the default) or 0/no/false, respectively.  See README.tasks for the
user-level details.

Note that there is a slight flaw in that if guard pages are enabled
the program will always print a message saying so.  My next commit
will put this message under the control of the QT_INFO environment
variable, like other qthreads internal messages.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@21328 3a8e244f-b0f2-452b-bcba-4c88e055c3ca