Implement task-private allocators.
This adds four function pointers to task-private data in the tasking
layer implementation (only fifo for now, the rest to follow) -- one
each for malloc, calloc, realloc and free.  These fields are set to
point to the memory-layer implementation, and the fields are copied
when a new task is created.

These fields may be overwritten in the wrapper function for an on
clause, to provide locale-specific memory management.  (That code has
not yet been supplied.)  The basic design would allow a Chapel program
to run without specifying any architecture (at least for the sake of
memory management).  It is up to the architectural description to set
up these pointers to point to its own allocator.




git-svn-id: http://svn.code.sf.net/p/chapel/code/branches/collaborations/hierarchical_locales@21331 3a8e244f-b0f2-452b-bcba-4c88e055c3ca