Make it easier to adjust the optimization level for SSCA2 compilations.
[Not reviewed; seemed simple enough not to need it.]

Replaced the explicit "--fast" in the compilation command line with the
expansion of the OPT Makefile variable.  The default value of OPT is
still "--fast", but it can be changed from the make(1) command line.

(This is one of a number of small test changes I've been carrying around
seemingly forever.)


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@21343 3a8e244f-b0f2-452b-bcba-4c88e055c3ca