Externalize streamCommCheck correctness testing.
Currently multilocale/deitz/needMultiLocales/streamCommCheck checks
its own comm counts internally, and effectively says "pass" or "fail"
itself.  There doesn't see to be an easy way to make it pass when it
is built with varying comm and/or tasking layer implementations which
result in different comm counts.  Right now it fails when built with
comm=ugni and tasks=muxed, because that combination results in 0
(zero) forks instead of the expected 1 (one).

This change makes streamCommCheck produce its comm counts explicitly,
so that the comparison with the "good" values can be sensitive to
which comm and tasking layers are being used.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@21344 3a8e244f-b0f2-452b-bcba-4c88e055c3ca