Support the launcher, and return to non-initialized allocations.
[Reviewed by Tom and Brad.]

In r21348 some of the allocations in chpltypes.c were changed from
calling chpl_mem_allocMany() to calling chpl_tracked_task_calloc().
Unfortunately this broke the launcher, because the task-private tracking
memory allocation functions aren't available there.  To get around this,
define a new allocator wrapper local to the file, which calls the task-
private tracking allocator when we're built for inclusion in the user
program, and calls the regular malloc() when we're built for inclusion
in the launcher.

In addition, r21348 inadvertently changed from doing malloc-like
non-initialized allocations to calloc-like initialized ones.  While
we're here, since the initialization isn't needed, undo that change.


git-svn-id: http://svn.code.sf.net/p/chapel/code/branches/collaborations/hierarchical_locales@21355 3a8e244f-b0f2-452b-bcba-4c88e055c3ca