Redefine chpl__initCopy(a) where a.type == string s.t. string_copy is not performed.
This patch also contains a number of futures required to get a clean standard
regression run.  With the string leaks tied up, it becomes obvious that the
remaining leaks are due to the nested var iterator bug.

Rationale: Strings in Chapel are const, so performing arg copies on them just
wastes memory.  The copies can be removed without changing semantics.  This was
proven extensively in the hierloc branch by disabling the copying and freeing
string in codegen for the last two weeks.   Local strings can be shared freely
(but we currently have no way to track their memory and delete them).

Caveat: This has been tested on the standard (--local) regressions.  Failures
are likely when --no-local or CHPL_COMM=gasnet are specified.

This is the main change: chpl__initCopy does not copy string anymore.  Removed
the code in codegen which avoided string_copy operations altogether.  In
callDestructors, removed the code which call field destructors on string fields.
All of these changes are temporary, pending David's reference-counted string
implementation.
  modules/internal/ChapelBase.chpl
  compiler/AST/expr.cpp
  compiler/resolution/callDestructors.cpp

Fixed the problem of the wrong descriptor message being printed for memory
allocated from Chapel code:
  runtime/include/chpl-mem-hook.h
  modules/internal/ChapelLocale.chpl

Futures added:
  test/arrays/deitz/part7/test_descriptor_frees.future
  test/memory/figueroa/LeakedMemory5.future
  test/memory/figueroa/LeakedMemoryArrayOfClasses.future
  test/memory/figueroa/LeakedMemory2.future

This test was restored to its former trunk splendor.
  test/memory/shannon/memmaxExceedsLineno2.chpl
  test/memory/shannon/memmaxExceedsLineno2.good
  test/memory/shannon/memmaxExceedsLineno2.future

Updated by the way.
  CHECKLIST-trunk-merge.txt

Some minor formatting changes.
  compiler/passes/normalize.cpp



git-svn-id: http://svn.code.sf.net/p/chapel/code/branches/collaborations/hierarchical_locales@21357 3a8e244f-b0f2-452b-bcba-4c88e055c3ca